requirements.txt
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.2.1
colorama==0.4.6
Flask==3.1.2
frozendict==2.4.6
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.2
multitasking==0.0.11
numpy==2.2.4
packaging==25.0
pandas==2.2.3
patsy==1.0.1
peewee==3.17.9
platformdirs==4.3.7
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytest==8.3.2
pytz==2025.2
requests==2.32.3
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
soupsieve==2.6
statsmodels==0.14.5
threadpoolctl==3.6.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
yfinance==0.2.55
pyinstaller
matplotlib
flet
pillow
